--- .vscode/launch.json
@@ -1,19 +1,0 @@
-{
-    // 使用 IntelliSense 了解相关属性。 
-    // 悬停以查看现有属性的描述。
-    // 欲了解更多信息，请访问: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "ours_next",
-            "request": "launch",
-            "type": "dart"
-        },
-        {
-            "name": "ours_next (profile mode)",
-            "request": "launch",
-            "type": "dart",
-            "flutterMode": "profile"
-        }
-    ]
-}
