--- lib/OursTheme.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/cupertino.dart';
-
-const Color OursColorMainRed = Color(0xFFBE4749);
-const Color OursColorStandOutRed = Color(0xFFE68E8F);
-const Color OursColorRedShadow = Color(0x61E68E8F);
-const Color OursColorBrightGray = Color(0xFFEEEEEE);
-const Color OursColorBrightStandOutRed = Color(0xFFF9DDDD);
-
-final BoxShadow normalShadow =
-    BoxShadow(color: OursColorRedShadow, offset: Offset(0, 46), blurRadius: 46);
-
-final BoxShadow normalShadow1 =
-    BoxShadow(color: OursColorRedShadow, offset: Offset(0, 4), blurRadius: 46);
